Add files via upload
ZIP file with MATLAB files.
The MATLAB scripts contain a Motion Energy Analysis (MEA) and scripts for the preprocessing of resulting time series. The MEA is frame differencing method for digital videos. Of freely selectable regions of interest the number of pixel is counted which color intensity change significant from t to t+1.